requirements.txt
click
chromadb
google-generativeai
python-dotenv